refactor style for SingupFormField
